Factor ElloUtil into Reachability and Alert
